Fix compilation errors from branch merge
- Fix DatasetPanel.cs: Remove non-existent DepthInterval property, fix PNM namespace, use ProjectManager for dataset operations
- Fix TextViewer.cs: Update export dialog API usage (Submit returns bool, use SelectedPath property)
- Fix NMRSimulationOpenCL.cs: Declare errorCode and device variables at method scope to avoid scoping issues
- Fix GeothermalOpenCLSolver.cs and BTESOpenCLSolver.cs: Cast ulong to long for Sum() to resolve ambiguous invocation

+++ Analysis/Geothermal/BTESOpenCLSolver.cs
@@ -222,7 +222,7 @@ public class BTESOpenCLSolver : IDisposable
 
                 DeviceName = string.Join(", ", deviceInfos.Select(d => d.Name));
                 DeviceVendor = string.Join(", ", deviceInfos.Select(d => d.Vendor).Distinct());
-                DeviceGlobalMemory = deviceInfos.Sum(d => d.GlobalMemory);
+                DeviceGlobalMemory = (ulong)deviceInfos.Sum(d => (long)d.GlobalMemory);
 
                 Logger.Log($"BTESOpenCLSolver: Multi-GPU mode with {_devices.Count} devices:");
                 foreach (var info in deviceInfos)
